Kaggle Amazon Employee Access Challenge
To take data about employee’s role information and a resource code and will return whether or not access should be granted or not.

The dataset consists of real historical data collected from 2010 & 2011. Each employee has certain information associated with them which is the criteria to decide whether a particular privilege to the employee will be given or not. 

The file contains EDA and different models approach.